Add Set nic as bridge master script
This script assumes enp3s0 is the ethernet device and creates bridge br0
The script requires an argument which is uses as the last part of the IP Address

On ak19 call the script with: sudo ~/qemu/set-nic-as-bridge-master.sh 19
On ak20 call the script with: sudo ~/qemu/set-nic-as-bridge-master.sh 20

The script will set the bridge IP Address to 10.55.5.$1/8
and a default route of 10.55.5.1 and remove the IP Address from the nic

This will allow me to launch virtual machines using QEMU that get an IP Address
on the host network.

This is all done with direct changes to the iptables and will not be retained
through a system reboot.

https://ak20.lam1.us/a/QEMU_(Quick_Emulator)_(software)