Add aws-lam.conf - For an instance with IPv4 access
This file defines aws, lam1, lam1a, lam2, lam2a and lam3 host entries along with
the aws-lam1, aws-lam2 and aws-lam3 tunnel definitions that allow access to an
IPv6 only instance via a ssh tunnel though the aws dual stack instance.

The aws, lam1 and lam2 enties are for access to a dual stack instance.

Delay the texlive install until after swap is enabled as a workaround to
resolv oom-kill of luahbtex and failure of "packages:" on VMEM=488 QEMU
instance cloud-config initialization.
This change allows a VMEM=488 QEMU instance to complete cloud-config
initialization without oom-kill interfering with the processes.

The texlive install with dependancies is 98 packages for Ubuntu 22.04

Omit hashing of new known host entries