Delay the texlive install until after swap
Delay the texlive install until after swap is enabled as a workaround to
allow a VMEM=488 QEMU instance to complete cloud-config initialization
without oom-kill interfering with the processes.

On openSUSE Leap 15.5 a VMEM=488 QEMU instance has 430Mi Mem an 0B Swap
before the swap partition is enabled in the runcmd section.
After that the instance has 430Mi Mem and 510Mi Swap via /dev/sdb1

The texlive install with dependancies is 180 packages for openSUSE